Send data to Fermentrack
