+++ src/wifi_setup.h
@@ -1,0 +1,12 @@
+//
+// Created by John Beeler on 6/4/18.
+//
+
+#ifndef TILTBRIDGE_WIFI_SETUP_H
+#define TILTBRIDGE_WIFI_SETUP_H
+
+
+void init_wifi();
+
+
+#endif //TILTBRIDGE_WIFI_SETUP_H
